[add] Look at all the pretty colors
