+++ src/ClearBurger.js
@@ -1,7 +1,8 @@
 import React from 'react'
 
 const ClearBurger = () => {
-<h1>Clear burger</h1>
+return(<h1>Clear burger</h1>
+)
 }
 
 export default ClearBurger
