+++ src/IngredientList.js
@@ -3,9 +3,13 @@ import Ingredients from './Ingredients';
 
 
 const IngredientList = (props) => {
+    
     return (
         <div className="ingredientList">
             <h1>Ingredient List</h1>
+            <ol>
+                <Ingredients />
+            </ol>
         </div>
     )
 }
